feat: refactored SwapForm into smaller components and hooks
